test(cli): testing cli emoji search :white_check_mark:
