+++ README.md
@@ -1,7 +1,7 @@
 # conventional-committer
 
 [![Build Status](https://travis-ci.com/wel-shy/committer.svg?branch=master)](https://travis-ci.com/wel-shy/committer)
-[![codecov](https://codecov.io/gh/wel-shy/committer/branch/master/graph/badge.svg)](https://codecov.io/gh/wel-shy/committer)
+[![Coverage Status](https://coveralls.io/repos/github/wel-shy/committer/badge.svg?branch=dev)](https://coveralls.io/github/wel-shy/committer?branch=dev)
 
 Git commit manager, format commits to meet the Conventional Commit format, but also include
 an emoji for enhanced visual grepping.
