+++ src/app.ts
@@ -10,105 +10,6 @@ import chalk from "chalk";
  */
 export default class App {
 
-  constructor() {
-  }
-
-  /**
-   * Get responses from command line
-   * @returns {Promise<{type: string; scope: string; description: string; body: string; footer: string; issue: string}>}
-   */
-  public async getAnswers(): Promise<{
-    type: string,
-    scope: string,
-    description: string,
-    body: string,
-    issue: string,
-    emoji: string,
-  }> {
-    const commitTypes = [
-      'build',
-      'ci',
-      'chore',
-      'docs',
-      'feat',
-      'fix',
-      'perf',
-      'refactor',
-      'revert',
-      'style',
-      'test'
-    ];
-
-    const questions = [
-      { type: 'autocomplete',
-        name: 'emoji',
-        message: 'Commit summary:',
-        /**
-         * Get source for emoji list.
-         * @param answersSoFar
-         * @param input
-         */
-        source: (answersSoFar: any, input: string) => {
-          return new Promise(resolve => {
-            // map to string
-            const mapped: string[] = emojis.gitmojis.map(gitmoji => `${gitmoji.emoji.trim()} - ${gitmoji.description.trim()}`);
-
-            // format input
-            if (input === undefined) {
-              input = '';
-            }
-            input = input.toLowerCase().trim();
-
-            // return matches
-            resolve(mapped.filter(gitmoji => gitmoji.toLowerCase().indexOf(input) > -1))
-          })
-        },
-        filter: (input: string) => {
-          return new Promise(resolve => {
-            const ans:
-            { "emoji": string,
-              "entity"?: string,
-              "code": string,
-              "description": string,
-              "name": string }[] = emojis.gitmojis.filter( g => input.indexOf(g.emoji) > -1);
-            resolve(ans[0].code);
-          })
-        }
-      },
-      { type: 'list', name: 'type', message: 'Choose commit type:', choices: commitTypes },
-      { type: 'input', name: 'scope', message: 'Enter commit scope (optional):'},
-      { type: 'input', name: 'description', message: 'Enter commit title:'},
-      { type: 'input', name: 'body', message: 'Enter commit body (optional):'},
-      { type: 'input', name: 'issue', message: 'References issue/PR (optional):'}
-    ];
-
-    // TODO: find a better way to do this with typescript.
-    inquirer.registerPrompt('autocomplete', require('inquirer-autocomplete-prompt'));
-    return await inquirer.prompt(questions);
-  }
-
-  /**
-   * Validate responses.
-   * @param {{type: string; scope: string; description: string; body: string; footer: string; issue: string; emoji: string}} answers
-   */
-  public validate(answers: {
-    type: string,
-    scope: string,
-    description: string,
-    body: string,
-    issue: string,
-    emoji: string,
-  }) {
-    if (answers.type === '') throw new Error('Commit must have a type');
-    if (answers.description === '') throw new Error('Commit must have a description');
-
-    if (answers.issue !== '') {
-      if (isNaN(parseInt(answers.issue))) {
-        throw new Error('Issue must be an integer');
-      }
-    }
-  }
-
   /**
    * Construct the commit message
    *
